+++ src/python/dxpy/sugar/chunking.py
@@ -13,6 +13,7 @@
 #   WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 #   License for the specific language governing permissions and limitations
 #   under the License.
+from __future__ import print_function, unicode_literals, division, absolute_import
 
 import dxpy
