+++ src/python/test/sugar/test_transfers.py
@@ -1,60 +1,13 @@
 from __future__ import print_function, unicode_literals, division, absolute_import
-import contextlib
 import gzip
 import os
-import shutil
 import subprocess
-import tempfile
 import unittest
-from uuid import uuid4
+
+from . import isolated_dir, make_random_files, random_name, run
 
 import dxpy
 from dxpy.sugar import transfers as xfer
-
-import logging
-logging.basicConfig(level="INFO")
-
-
-def random_name(fmt_str=None):
-    rndstr = str(uuid4())
-    if fmt_str:
-        return fmt_str.format(rndstr)
-    else:
-        return rndstr
-
-
-def make_random_files(n, fmt_str=None):
-    filenames = []
-    for i in range(n):
-        fname = random_name(fmt_str)
-        filenames.append(fname)
-        with open(fname, "wt") as out:
-            out.write("test{}".format(i))
-    return filenames
-
-
-@contextlib.contextmanager
-def temp_dir(*args, **kwargs):
-    dname = tempfile.mkdtemp(*args, **kwargs)
-    try:
-        yield dname
-    finally:
-        shutil.rmtree(dname)
-
-
-@contextlib.contextmanager
-def isolated_dir():
-    with temp_dir() as d:
-        curdir = os.getcwd()
-        os.chdir(d)
-        try:
-            yield d
-        finally:
-            os.chdir(curdir)
-
-
-def run(cmd):
-    subprocess.check_call(cmd, shell=True)
 
 
 class TestUpload(unittest.TestCase):
